[#185] 오디오플레이어 container 분리
